pyproject.toml
[tool.poetry]
include = [
    "aiconsole/template/**/*",
    "aiconsole/static/**/*"
]
name = "aiconsole"
version = "0.2.0-alpha.1"
description = "Run and expand your personal AI tools"
authors = ["Maciej Cielecki <maciej.cielecki@10clouds.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tiktoken = "^0.5.1"
openai-function-call = "^0.2.6"
watchdog = "^3.0.0"
litellm = "^0.6.2"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
beautifulsoup4 = "^4.12.2"
selenium = "^4.14.0"
webdriver-manager = "^4.0.1"
pillow = "^10.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aiconsole = "aiconsole.init:aiconsole"
aiconsole-dev = "aiconsole.init:aiconsole_dev"
